[bugfix] variant sites may be skipped
If the BAM file has reads from contigs/chromosomes prior to the first variant in the VCF, but those reads have position greater than the first variant in the VCF, then the first variant may be skipped.  For example, if the BAM file has a read from `chr3:100-200` while the VCF has a variant at `chr4:50`.  The variant is latter in the genome by coordinate, but if we only compare the position (i.e. 100 for the read and 50 for the variant), it is "earlier". 

The issue stemmed from the fact that read's contig was not used when clearing the variant buffer, but instead the variant itself!